feat(video-player): implement web video player tests and stubs
