+++ mobile/packages/divine_video_player/test/src/html_video_element_backend_web_test_stub.dart
@@ -1,0 +1,1 @@
+void main() {}
